// app/lib/types.ts
export type AreaType = 'region' | 'nhsRegion';

export type RegionAreaName =
  | 'East Midlands'
  | 'East of England'
  | 'London'
  | 'North East'
  | 'North West'
  | 'South East'
  | 'South West'
  | 'West Midlands'
  | 'Yorkshire and The Humber';

export type NHSAreaName =
  | 'Midlands'
  | 'East of England'
  | 'London'
  | 'North East and Yorkshire'
  | 'North West'
  | 'South East'
  | 'South West';

export type ChangeDirection = 'UP' | 'DOWN' | 'SAME' | 'NONE';